@@ -35,6 +35,7 @@ flutter:
         default_package: mapbox_gl_web
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  # Flutter versions prior to 1.10 did not support the flutter.plugin.platforms map.
-  flutter: ">=2.0.0"
+  # Flutter 2.15 is the required min to be able to use Enum.name
+  sdk: '>=2.15.0 <3.0.0'
+  flutter: ">=2.8.1"
+
