@@ -9,7 +9,7 @@ import 'page.dart';
 import 'util.dart';
 
 class ClickAnnotationPage extends ExamplePage {
-  ClickAnnotationPage()
+  const ClickAnnotationPage({super.key})
       : super(const Icon(Icons.check_circle), 'Annotation tap');
 
   @override
@@ -19,7 +19,7 @@ class ClickAnnotationPage extends ExamplePage {
 }
 
 class ClickAnnotationBody extends StatefulWidget {
-  const ClickAnnotationBody();
+  const ClickAnnotationBody({super.key});
 
   @override
   State<StatefulWidget> createState() => ClickAnnotationBodyState();
@@ -27,7 +27,8 @@ class ClickAnnotationBody extends StatefulWidget {
 
 class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
   ClickAnnotationBodyState();
-  static const LatLng center = const LatLng(-33.88, 151.16);
+
+  static const LatLng center = LatLng(-33.88, 151.16);
 
   MaplibreMapController? controller;
 
@@ -50,8 +51,13 @@ class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
 
   _showSnackBar(String type, String id) {
     final snackBar = SnackBar(
-        content: Text('Tapped $type $id',
-            style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold)),
+        content: Text(
+          'Tapped $type $id',
+          style: const TextStyle(
+            fontSize: 18,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
         backgroundColor: Theme.of(context).primaryColor);
     ScaffoldMessenger.of(context).clearSnackBars();
     ScaffoldMessenger.of(context).showSnackBar(snackBar);
@@ -77,7 +83,7 @@ class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
     await addImageFromAsset(
         controller!, "custom-marker", "assets/symbols/custom-marker.png");
     controller!.addCircle(
-      CircleOptions(
+      const CircleOptions(
         geometry: LatLng(-33.881979408447314, 151.171361438502117),
         circleStrokeColor: "#00FF00",
         circleStrokeWidth: 2,
@@ -85,7 +91,7 @@ class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
       ),
     );
     controller!.addCircle(
-      CircleOptions(
+      const CircleOptions(
         geometry: LatLng(-33.894372606072309, 151.17576679759523),
         circleStrokeColor: "#00FF00",
         circleStrokeWidth: 2,
@@ -93,13 +99,13 @@ class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
       ),
     );
     controller!.addSymbol(
-      SymbolOptions(
+      const SymbolOptions(
           geometry: LatLng(-33.894372606072309, 151.17576679759523),
           iconImage: "custom-marker", //"fast-food-15",
           iconSize: 2),
     );
     controller!.addLine(
-      LineOptions(
+      const LineOptions(
         geometry: [
           LatLng(-33.874867744475786, 151.170627211986584),
           LatLng(-33.881979408447314, 151.171361438502117),
@@ -113,7 +119,7 @@ class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
     );
 
     controller!.addFill(
-      FillOptions(
+      const FillOptions(
         geometry: [
           [
             LatLng(-33.901517742631846, 151.178099204457737),
@@ -134,7 +140,7 @@ class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
   @override
   Widget build(BuildContext context) {
     return MaplibreMap(
-      annotationOrder: [
+      annotationOrder: const [
         AnnotationType.fill,
         AnnotationType.line,
         AnnotationType.circle,
