@@ -11,7 +11,7 @@ import 'package:maplibre_gl/mapbox_gl.dart';
 import 'page.dart';
 
 class LinePage extends ExamplePage {
-  LinePage() : super(const Icon(Icons.share), 'Line');
+  const LinePage({super.key}) : super(const Icon(Icons.share), 'Line');
 
   @override
   Widget build(BuildContext context) {
@@ -20,7 +20,7 @@ class LinePage extends ExamplePage {
 }
 
 class LineBody extends StatefulWidget {
-  const LineBody();
+  const LineBody({super.key});
 
   @override
   State<StatefulWidget> createState() => LineBodyState();
@@ -29,7 +29,7 @@ class LineBody extends StatefulWidget {
 class LineBodyState extends State<LineBody> {
   LineBodyState();
 
-  static final LatLng center = const LatLng(-33.86711, 151.1947171);
+  static const LatLng center = LatLng(-33.86711, 151.1947171);
 
   MaplibreMapController? controller;
   int _lineCount = 0;
@@ -56,13 +56,13 @@ class LineBodyState extends State<LineBody> {
 
   _onLineTapped(Line line) async {
     await _updateSelectedLine(
-      LineOptions(lineColor: "#ff0000"),
+      const LineOptions(lineColor: "#ff0000"),
     );
     setState(() {
       _selectedLine = line;
     });
     await _updateSelectedLine(
-      LineOptions(lineColor: "#ffe100"),
+      const LineOptions(lineColor: "#ffe100"),
     );
   }
 
@@ -72,7 +72,7 @@ class LineBodyState extends State<LineBody> {
 
   void _add() {
     controller!.addLine(
-      LineOptions(
+      const LineOptions(
           geometry: [
             LatLng(-33.86711, 151.1947171),
             LatLng(-33.86711, 151.1947171),
@@ -118,10 +118,7 @@ class LineBodyState extends State<LineBody> {
 
   Future<void> _changeAlpha() async {
     double? current = _selectedLine!.options.lineOpacity;
-    if (current == null) {
-      // default value
-      current = 1.0;
-    }
+    current ??= 1.0;
 
     await _updateSelectedLine(
       LineOptions(lineOpacity: current < 0.1 ? 1.0 : current * 0.75),
@@ -130,10 +127,7 @@ class LineBodyState extends State<LineBody> {
 
   Future<void> _toggleVisible() async {
     double? current = _selectedLine!.options.lineOpacity;
-    if (current == null) {
-      // default value
-      current = 1.0;
-    }
+    current ??= 1.0;
     await _updateSelectedLine(
       LineOptions(lineOpacity: current == 0.0 ? 1.0 : 0.0),
     );
@@ -142,7 +136,7 @@ class LineBodyState extends State<LineBody> {
   _onStyleLoadedCallback() async {
     addImageFromAsset("assetImage", _linePatternImage);
     await controller!.addLine(
-      LineOptions(
+      const LineOptions(
         geometry: [LatLng(37.4220, -122.0841), LatLng(37.4240, -122.0941)],
         lineColor: "#ff0000",
         lineWidth: 14.0,
@@ -180,52 +174,52 @@ class LineBodyState extends State<LineBody> {
                     Row(
                       children: <Widget>[
                         TextButton(
-                          child: const Text('add'),
                           onPressed: (_lineCount == 12) ? null : _add,
+                          child: const Text('add'),
                         ),
                         TextButton(
-                          child: const Text('remove'),
                           onPressed: (_selectedLine == null) ? null : _remove,
+                          child: const Text('remove'),
                         ),
                         TextButton(
-                          child: const Text('move'),
                           onPressed: (_selectedLine == null)
                               ? null
                               : () async {
                                   await _move();
                                 },
+                          child: const Text('move'),
                         ),
                         TextButton(
-                          child: const Text('change line-pattern'),
                           onPressed: (_selectedLine == null)
                               ? null
                               : _changeLinePattern,
+                          child: const Text('change line-pattern'),
                         ),
                       ],
                     ),
                     Row(
                       children: <Widget>[
                         TextButton(
-                          child: const Text('change alpha'),
                           onPressed:
                               (_selectedLine == null) ? null : _changeAlpha,
+                          child: const Text('change alpha'),
                         ),
                         TextButton(
-                          child: const Text('toggle visible'),
                           onPressed:
                               (_selectedLine == null) ? null : _toggleVisible,
+                          child: const Text('toggle visible'),
                         ),
                         TextButton(
-                          child: const Text('print current LatLng'),
                           onPressed: (_selectedLine == null)
                               ? null
                               : () async {
                                   var latLngs = await controller!
                                       .getLineLatLngs(_selectedLine!);
                                   for (var latLng in latLngs) {
-                                    print(latLng.toString());
+                                    debugPrint(latLng.toString());
                                   }
                                 },
+                          child: const Text('print current LatLng'),
                         ),
                       ],
                     ),
