@@ -98,7 +98,8 @@ final List<OfflineRegionListItem> allRegions = [
 ];
 
 class OfflineRegionsPage extends ExamplePage {
-  OfflineRegionsPage() : super(const Icon(Icons.map), 'Offline Regions');
+  const OfflineRegionsPage({super.key})
+      : super(const Icon(Icons.map), 'Offline Regions');
 
   @override
   Widget build(BuildContext context) {
@@ -107,14 +108,14 @@ class OfflineRegionsPage extends ExamplePage {
 }
 
 class OfflineRegionBody extends StatefulWidget {
-  const OfflineRegionBody();
+  const OfflineRegionBody({super.key});
 
   @override
-  _OfflineRegionsBodyState createState() => _OfflineRegionsBodyState();
+  State<OfflineRegionBody> createState() => _OfflineRegionsBodyState();
 }
 
 class _OfflineRegionsBodyState extends State<OfflineRegionBody> {
-  List<OfflineRegionListItem> _items = [];
+  final List<OfflineRegionListItem> _items = [];
 
   @override
   void initState() {
@@ -133,7 +134,7 @@ class _OfflineRegionsBodyState extends State<OfflineRegionBody> {
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
               IconButton(
-                icon: Icon(Icons.map),
+                icon: const Icon(Icons.map),
                 onPressed: () => _goToMap(_items[index]),
               ),
               Column(
@@ -142,14 +143,14 @@ class _OfflineRegionsBodyState extends State<OfflineRegionBody> {
                 children: <Widget>[
                   Text(
                     _items[index].name,
-                    style: TextStyle(
+                    style: const TextStyle(
                       fontWeight: FontWeight.bold,
                       fontSize: 18,
                     ),
                   ),
                   Text(
                     'Est. tiles: ${_items[index].estimatedTiles}',
-                    style: TextStyle(
+                    style: const TextStyle(
                       fontSize: 16,
                     ),
                   ),
@@ -157,10 +158,10 @@ class _OfflineRegionsBodyState extends State<OfflineRegionBody> {
               ),
               const Spacer(),
               _items[index].isDownloading
-                  ? Container(
-                      child: CircularProgressIndicator(),
+                  ? const SizedBox(
                       height: 16,
                       width: 16,
+                      child: CircularProgressIndicator(),
                     )
                   : IconButton(
                       icon: Icon(
