@@ -45,7 +45,8 @@ const fillOptions = [
 ];
 
 class BatchAddPage extends ExamplePage {
-  BatchAddPage() : super(const Icon(Icons.check_circle), 'Batch add/remove');
+  const BatchAddPage({super.key})
+      : super(const Icon(Icons.check_circle), 'Batch add/remove');
 
   @override
   Widget build(BuildContext context) {
@@ -54,7 +55,7 @@ class BatchAddPage extends ExamplePage {
 }
 
 class BatchAddBody extends StatefulWidget {
-  const BatchAddBody();
+  const BatchAddBody({super.key});
 
   @override
   State<StatefulWidget> createState() => BatchAddBodyState();
@@ -62,12 +63,13 @@ class BatchAddBody extends StatefulWidget {
 
 class BatchAddBodyState extends State<BatchAddBody> {
   BatchAddBodyState();
+
   List<Fill> _fills = [];
   List<Circle> _circles = [];
   List<Line> _lines = [];
   List<Symbol> _symbols = [];
 
-  static final LatLng center = const LatLng(-33.86711, 151.1947171);
+  static const LatLng center = LatLng(-33.86711, 151.1947171);
 
   late MaplibreMapController controller;
 
@@ -104,11 +106,12 @@ class BatchAddBodyState extends State<BatchAddBody> {
     final symbolOptions = <SymbolOptions>[];
     for (final option in options) {
       // put symbols only on the inner most ring if it exists
-      if (option.geometry!.length > 1)
+      if (option.geometry!.length > 1) {
         for (final latLng in option.geometry!.last) {
           symbolOptions
               .add(SymbolOptions(iconImage: 'custom-marker', geometry: latLng));
         }
+      }
     }
     return symbolOptions;
   }
@@ -172,10 +175,10 @@ class BatchAddBodyState extends State<BatchAddBody> {
                     Column(
                       children: <Widget>[
                         TextButton(
-                            child: const Text('batch add'), onPressed: _add),
+                            onPressed: _add, child: const Text('batch add')),
                         TextButton(
-                            child: const Text('batch remove'),
-                            onPressed: _remove),
+                            onPressed: _remove,
+                            child: const Text('batch remove')),
                       ],
                     ),
                   ],
