@@ -11,7 +11,8 @@ import 'package:maplibre_gl/mapbox_gl.dart';
 import 'page.dart';
 
 class PlaceFillPage extends ExamplePage {
-  PlaceFillPage() : super(const Icon(Icons.check_circle), 'Place fill');
+  const PlaceFillPage({super.key})
+      : super(const Icon(Icons.check_circle), 'Place fill');
 
   @override
   Widget build(BuildContext context) {
@@ -20,7 +21,7 @@ class PlaceFillPage extends ExamplePage {
 }
 
 class PlaceFillBody extends StatefulWidget {
-  const PlaceFillBody();
+  const PlaceFillBody({super.key});
 
   @override
   State<StatefulWidget> createState() => PlaceFillBodyState();
@@ -29,22 +30,22 @@ class PlaceFillBody extends StatefulWidget {
 class PlaceFillBodyState extends State<PlaceFillBody> {
   PlaceFillBodyState();
 
-  static final LatLng center = const LatLng(-33.86711, 151.1947171);
+  static const LatLng center = LatLng(-33.86711, 151.1947171);
   final String _fillPatternImage = "assets/fill/cat_silhouette_pattern.png";
 
   final List<List<LatLng>> _defaultGeometry = [
     [
-      LatLng(-33.719, 151.150),
-      LatLng(-33.858, 151.150),
-      LatLng(-33.866, 151.401),
-      LatLng(-33.747, 151.328),
-      LatLng(-33.719, 151.150),
+      const LatLng(-33.719, 151.150),
+      const LatLng(-33.858, 151.150),
+      const LatLng(-33.866, 151.401),
+      const LatLng(-33.747, 151.328),
+      const LatLng(-33.719, 151.150),
     ],
     [
-      LatLng(-33.762, 151.250),
-      LatLng(-33.827, 151.250),
-      LatLng(-33.833, 151.347),
-      LatLng(-33.762, 151.250),
+      const LatLng(-33.762, 151.250),
+      const LatLng(-33.827, 151.250),
+      const LatLng(-33.833, 151.347),
+      const LatLng(-33.762, 151.250),
     ]
   ];
 
@@ -128,9 +129,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   void _changePosition() {
     List<List<LatLng>>? geometry = _selectedFill!.options.geometry;
 
-    if (geometry == null) {
-      geometry = _defaultGeometry;
-    }
+    geometry ??= _defaultGeometry;
 
     _updateSelectedFill(FillOptions(
         geometry: geometry
@@ -144,10 +143,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
 
   void _changeDraggable() {
     bool? draggable = _selectedFill!.options.draggable;
-    if (draggable == null) {
-      // default value
-      draggable = false;
-    }
+    draggable ??= false;
     _updateSelectedFill(
       FillOptions(draggable: !draggable),
     );
@@ -155,10 +151,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
 
   Future<void> _changeFillOpacity() async {
     double? current = _selectedFill!.options.fillOpacity;
-    if (current == null) {
-      // default value
-      current = 1.0;
-    }
+    current ??= 1.0;
 
     _updateSelectedFill(
       FillOptions(fillOpacity: current < 0.1 ? 1.0 : current * 0.75),
@@ -167,25 +160,19 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
 
   Future<void> _changeFillColor() async {
     String? current = _selectedFill!.options.fillColor;
-    if (current == null) {
-      // default value
-      current = "#FF0000";
-    }
+    current ??= "#FF0000";
 
     _updateSelectedFill(
-      FillOptions(fillColor: "#FFFF00"),
+      const FillOptions(fillColor: "#FFFF00"),
     );
   }
 
   Future<void> _changeFillOutlineColor() async {
     String? current = _selectedFill!.options.fillOutlineColor;
-    if (current == null) {
-      // default value
-      current = "#FF0000";
-    }
+    current ??= "#FF0000";
 
     _updateSelectedFill(
-      FillOptions(fillOutlineColor: "#FFFF00"),
+      const FillOptions(fillOutlineColor: "#FFFF00"),
     );
   }
 
@@ -227,49 +214,49 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
                     Column(
                       children: <Widget>[
                         TextButton(
-                          child: const Text('add'),
                           onPressed: (_fillCount == 12) ? null : _add,
+                          child: const Text('add'),
                         ),
                         TextButton(
-                          child: const Text('remove'),
                           onPressed: (_selectedFill == null) ? null : _remove,
+                          child: const Text('remove'),
                         ),
                       ],
                     ),
                     Column(
                       children: <Widget>[
                         TextButton(
-                          child: const Text('change fill-opacity'),
                           onPressed: (_selectedFill == null)
                               ? null
                               : _changeFillOpacity,
+                          child: const Text('change fill-opacity'),
                         ),
                         TextButton(
-                          child: const Text('change fill-color'),
                           onPressed:
                               (_selectedFill == null) ? null : _changeFillColor,
+                          child: const Text('change fill-color'),
                         ),
                         TextButton(
-                          child: const Text('change fill-outline-color'),
                           onPressed: (_selectedFill == null)
                               ? null
                               : _changeFillOutlineColor,
+                          child: const Text('change fill-outline-color'),
                         ),
                         TextButton(
-                          child: const Text('change fill-pattern'),
                           onPressed: (_selectedFill == null)
                               ? null
                               : _changeFillPattern,
+                          child: const Text('change fill-pattern'),
                         ),
                         TextButton(
-                          child: const Text('change position'),
                           onPressed:
                               (_selectedFill == null) ? null : _changePosition,
+                          child: const Text('change position'),
                         ),
                         TextButton(
-                          child: const Text('toggle draggable'),
                           onPressed:
                               (_selectedFill == null) ? null : _changeDraggable,
+                          child: const Text('toggle draggable'),
                         ),
                       ],
                     ),
