@@ -11,7 +11,8 @@ import 'package:maplibre_gl/mapbox_gl.dart';
 import 'page.dart';
 
 class PlaceSourcePage extends ExamplePage {
-  PlaceSourcePage() : super(const Icon(Icons.place), 'Place source');
+  const PlaceSourcePage({super.key})
+      : super(const Icon(Icons.place), 'Place source');
 
   @override
   Widget build(BuildContext context) {
@@ -20,7 +21,7 @@ class PlaceSourcePage extends ExamplePage {
 }
 
 class PlaceSymbolBody extends StatefulWidget {
-  const PlaceSymbolBody();
+  const PlaceSymbolBody({super.key});
 
   @override
   State<StatefulWidget> createState() => PlaceSymbolBodyState();
@@ -29,8 +30,8 @@ class PlaceSymbolBody extends StatefulWidget {
 class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   PlaceSymbolBodyState();
 
-  static const SOURCE_ID = 'sydney_source';
-  static const LAYER_ID = 'sydney_layer';
+  static const sourceId = 'sydney_source';
+  static const layerId = 'sydney_layer';
 
   bool sourceAdded = false;
   bool layerAdded = false;
@@ -116,44 +117,44 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
                 Column(
                   children: <Widget>[
                     TextButton(
-                      child: const Text('Add source (asset image)'),
                       onPressed: sourceAdded
                           ? null
                           : () {
                               addImageSourceFromAsset(
-                                      SOURCE_ID, 'assets/sydney.png')
+                                      sourceId, 'assets/sydney.png')
                                   .then((value) {
                                 setState(() => sourceAdded = true);
                               });
                             },
+                      child: const Text('Add source (asset image)'),
                     ),
                     TextButton(
-                      child: const Text('Remove source (asset image)'),
                       onPressed: sourceAdded
                           ? () async {
-                              await removeLayer(LAYER_ID);
-                              removeImageSource(SOURCE_ID).then((value) {
+                              await removeLayer(layerId);
+                              removeImageSource(sourceId).then((value) {
                                 setState(() => sourceAdded = false);
                               });
                             }
                           : null,
+                      child: const Text('Remove source (asset image)'),
                     ),
                     TextButton(
+                      onPressed: sourceAdded
+                          ? () => addLayer(layerId, sourceId)
+                          : null,
                       child: const Text('Show layer'),
-                      onPressed: sourceAdded
-                          ? () => addLayer(LAYER_ID, SOURCE_ID)
-                          : null,
                     ),
                     TextButton(
+                      onPressed: sourceAdded
+                          ? () => addLayerBelow(layerId, sourceId, 'water')
+                          : null,
                       child: const Text('Show layer below water'),
-                      onPressed: sourceAdded
-                          ? () => addLayerBelow(LAYER_ID, SOURCE_ID, 'water')
-                          : null,
                     ),
                     TextButton(
-                      child: const Text('Hide layer'),
                       onPressed:
-                          sourceAdded ? () => removeLayer(LAYER_ID) : null,
+                          sourceAdded ? () => removeLayer(layerId) : null,
+                      child: const Text('Hide layer'),
                     ),
                   ],
                 ),
