@@ -28,7 +28,7 @@ class MaplibreMap extends StatefulWidget {
     this.trackCameraPosition = false,
     this.myLocationEnabled = false,
     this.myLocationTrackingMode = MyLocationTrackingMode.None,
-    this.myLocationRenderMode = MyLocationRenderMode.COMPASS,
+    this.myLocationRenderMode = MyLocationRenderMode.NORMAL,
     this.logoViewMargins,
     this.compassViewPosition,
     this.compassViewMargins,
@@ -53,9 +53,13 @@ class MaplibreMap extends StatefulWidget {
       AnnotationType.line,
       AnnotationType.circle,
     ],
-    this.useDelayedDisposal,
-    this.useHybridCompositionOverride,
-  })  : assert(annotationOrder.length <= 4),
+  })  : assert(
+          myLocationRenderMode != MyLocationRenderMode.NORMAL
+              ? myLocationEnabled
+              : true,
+          "$myLocationRenderMode requires [myLocationEnabled] set to true.",
+        ),
+        assert(annotationOrder.length <= 4),
         assert(annotationConsumeTapEvents.length > 0),
         super(key: key);
 
@@ -157,7 +161,9 @@ class MaplibreMap extends StatefulWidget {
   /// `myLocationEnabled` needs to be true for values other than `MyLocationTrackingMode.None` to work.
   final MyLocationTrackingMode myLocationTrackingMode;
 
-  /// The mode to render the user location symbol
+  /// Specifies if and how the user's heading/bearing is rendered in the user location indicator.
+  /// See the documentation of [MyLocationRenderMode] for details.
+  /// If this is set to a value other than [MyLocationRenderMode.NORMAL], [myLocationEnabled] needs to be true.
   final MyLocationRenderMode myLocationRenderMode;
 
   /// Set the layout margins for the Mapbox Logo
@@ -213,12 +219,6 @@ class MaplibreMap extends StatefulWidget {
   /// * All fade/transition animations have completed
   final OnMapIdleCallback? onMapIdle;
 
-  /// Use delayed disposal of Android View Controller to avoid flutter 3.x.x crashes
-  final bool? useDelayedDisposal;
-
-  /// Override hybrid mode per map instance
-  final bool? useHybridCompositionOverride;
-
   /// Set `MapboxMap.useHybridComposition` to `false` in order use Virtual-Display
   /// (better for Android 9 and below but may result in errors on Android 12)
   /// or leave it `true` (default) to use Hybrid composition (Slower on Android 9 and below).
@@ -250,8 +250,6 @@ class _MaplibreMapState extends State<MaplibreMap> {
       'options': _MapboxMapOptions.fromWidget(widget).toMap(),
       //'onAttributionClickOverride': widget.onAttributionClick != null,
       'dragEnabled': widget.dragEnabled,
-      'useDelayedDisposal': widget.useDelayedDisposal,
-      'useHybridCompositionOverride': widget.useHybridCompositionOverride,
     };
     return _mapboxGlPlatform.buildView(
         creationParams, onPlatformViewCreated, widget.gestureRecognizers);
